LevelHandler commit 4
Final commit for tonight.

Reflecting on this, there is probably a better way to do things then whatever I taped together these past few hours.

Feel free to change things as you please.

I have also added the testLevelFile to the repo, mostly so you guys know what I was dealing with when developing this and what changes were made to the format to get this 'far' - I will elaborate on what changes were made on the discord.